integrate new endpoint in client
