+++ packages/store-indexer/src/postgres/deprecated/createQueryAdapter.ts
@@ -3,7 +3,7 @@ import { PgDatabase } from "drizzle-orm/pg-core";
 import { TableWithRecords, isTableRegistrationLog, logToTable, storeTables } from "@latticexyz/store-sync";
 import { decodeKey, decodeValueArgs } from "@latticexyz/protocol-parser";
 import { QueryAdapter } from "@latticexyz/store-sync/trpc-indexer";
-import { debug } from "../debug";
+import { debug } from "../../debug";
 import { getLogs } from "./getLogs";
 import { groupBy } from "@latticexyz/common/utils";
 
@@ -12,6 +12,7 @@ import { groupBy } from "@latticexyz/common/utils";
  *
  * @param {PgDatabase<any>} database Postgres database object from Drizzle
  * @returns {Promise<QueryAdapter>} A set of methods used by tRPC endpoints.
+ * @deprecated
  */
 export async function createQueryAdapter(database: PgDatabase<any>): Promise<QueryAdapter> {
   const adapter: QueryAdapter = {
