+++ packages/store-sync/src/indexer-client/index.ts
@@ -1,0 +1,1 @@
+export * from "./createIndexerClient";
